broke down the keys module into smaller files
